requirements.txt
langchain
ipykernel
python-dotenv
langchain_community
pypdf
arxiv
bs4
langchain-text-splitters
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
fastapi
langchain_groq
uvicorn
sse_starlette
langserve
validators==0.28.1
SQLAlchemy
youtube_transcript_api
pytube
unstructured
gradio
langchain_astradb